requirements.txt
mplayer.py==0.7.1
